Create Mobsim Send SMS
Communicate with your customers in an assertive and direct way using the "MOBSIM - Send SMS" Connector. SMS - with high delivery and opening rates - makes it possible to send promotional and transactional campaigns, alerts, reminders, notifications, invites and more, allowing for a complete experience with personalized messages and the convenience of scheduled deliveries to mobile phone users.